Change MaskGIT name Add missing files
