+++ MaskGIT/.gitignore
@@ -1,0 +1,27 @@
+models/__pycache__/
+
+*.pyc
+
+VQGAN/pretrain/*
+!VQGAN/pretrain/README.md
+
+Data/*
+!Data/README.md
+
+
+SavedModels/*
+!SavedModels/README.md
+
+
+taming/*
+
+Output/Images
+
+FID_Images/Images
+
+figs/
+
+
+.ipynb_checkpoints/
+
+__pycache__/
