+++ MaskGIT/VQGAN/pretrain/README.md
@@ -1,0 +1,1 @@
+Pretrained VQGAN model comes here!
