+++ MaskGIT/info.yaml
@@ -1,0 +1,6 @@
+- acronym: MaskGIT 
+  title: "MaskGIT: Masked Generative Image Transformer"
+  year: 2022
+  venue: CVPR2022
+  paper_authors: Huiwen Chang, Han Zhang, Lu Jiang, Ce Liu, William T. Freeman
+  group_members: Ardan Yılmaz, Şahin Umutcan Üstündaş
