--- maskgit-ceng796/VQGAN/pretrain/README.md
@@ -1,1 +1,0 @@
-Pretrained VQGAN model comes here!
